backend/requirements.txt
alembic==1.14.1

Authlib==1.3.2
blinker==1.8.2
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
distlib==0.4.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0

Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
google-auth==2.40.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10

MarkupSafe==2.1.5

oauthlib==2.1.0
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2

pycparser==2.22

PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.4
requests-oauthlib==1.1.0
rsa==4.9.1
SQLAlchemy==2.0.41

typing_extensions==4.13.2
urllib3==2.2.3
virtualenv==20.31.2
Werkzeug==3.0.6

zipp==3.20.2
